requirements.txt
python-telegram-bot==21.6
openai>=1.54.0
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1